move models/sprites/sounds to an hlcoop subfolder
This will keep mod files separate from vanilla HL files when running as a Half-Life server. Files that override that default files will go here, and updates to files will need _v2 suffixes or similar, because servers will be responsible for sending mod updates.

The idea is that no one is going to download and play a mod like this when Sven Co-op exists, but there are currently thousands of players on vanilla HL thanks to the anniversary update. By making the mod compatible with vanilla HL servers at least some people will see this mod and join.

So, this mod will now function more like an elaborate plugin and not a standalone mod.