add skybox bsp template for trigger_changesky
Used to convert sky textures to a model that can be rendered in front of the normal sky. A BSP was used instead of a MDL because the world lighting doesn't affect BSPs when used as entity models, and the entity will render everywhere regardless of VIS. This BSP weighs ~130kb more than a model would due to mipmaps. There is less color depth than normal skies, and the linear texture filter mode is not forced for this kind of sky.

It was created by compiling a huge 65k cube, then scaled up 16 times with glitch-life to get around the RAD compiler limits, then cleaned with bspguy to delete some extra light data. Each face size is 1,048,576 units, which should not noticeably move around even in the largest sven maps.