add fog model and sources
This is the hackiest solution for fog ever. The model scale is controlled by animations, and color by bodies which map to a single pixel in a swappable texture. An offset is applied so that the model can be lit with EF_BRIGHTLIGHT outside of the map somewhere. Add a bunch of these models surrounding the player at different scales and renderamts to create "fog".