reduce bits for beam positions
these don't need fine precision. On svencoop2 you can see the beam chasing Tor has a little jitter with and without this change. If it can't be perfect then why bother with the extra bits. Most beams are chaotic and it wouldn't matter if they're off by even a few units. Weapon beams are client-sided and aren't affected by this, which is where you might notice the loss in precision.